software/doc/requirements.txt
Sphinx
sphinx-rtd-theme
sphinxcontrib-plantuml
